# cmd/api/main.go
package main

import (
	"log"
	"net/http"
	"strings"
	"time"

	"github.com/ahmadeko2017/backed-golang-tugas/internal/handler"
	"github.com/ahmadeko2017/backed-golang-tugas/internal/middleware"
	"github.com/ahmadeko2017/backed-golang-tugas/internal/repository"
	"github.com/ahmadeko2017/backed-golang-tugas/internal/service"
	"github.com/ahmadeko2017/backed-golang-tugas/pkg/config"
	"github.com/ahmadeko2017/backed-golang-tugas/pkg/database"
	"github.com/gin-gonic/gin"

	_ "github.com/ahmadeko2017/backed-golang-tugas/docs" // This will be generated by swag init
)

// @title           Category API
// @version         1.0
// @BasePath  /

func main() {
	// Load configuration (env, optional config file)
	config.Load()

	// Connect to DB
	database.Connect()

	// Setup Repository
	categoryRepo := repository.NewCategoryRepository(database.DB)

	// Setup Service
	categoryService := service.NewCategoryService(categoryRepo)

	// Setup Handler
	categoryHandler := handler.NewCategoryHandler(categoryService)

	// Setup Product Dependencies
	productRepo := repository.NewProductRepository(database.DB)
	productService := service.NewProductService(productRepo, categoryRepo)
	productHandler := handler.NewProductHandler(productService)
	// Setup Transaction Dependencies
	transactionRepo := repository.NewTransactionRepository(database.DB)
	transactionService := service.NewTransactionService(transactionRepo, productRepo)
	transactionHandler := handler.NewTransactionHandler(transactionService)

	// Setup Router
	r := gin.New()
	r.Use(gin.Logger())
	r.Use(middleware.RecoveryMiddleware())

	// Security Headers
	r.Use(middleware.SecurityHeadersMiddleware())

	// Body Size Limit (1MB)
	r.Use(middleware.BodySizeMiddleware(1024 * 1024))

	// CORS
	r.Use(middleware.CORSMiddleware())

	// Rate Limit
	r.Use(middleware.RateLimitMiddleware(10))

	// Health Check
	startTime := time.Now()
	healthHandler := handler.NewHealthHandler(startTime)
	r.GET("/health", healthHandler.HealthCheck)

	// Root Endpoint - Welcome & Redirect info
	r.Static("/static", "./web")
	r.GET("/", func(c *gin.Context) {
		c.File("./web/index.html")
	})

	// Swagger - Custom HTML to hide topbar
	r.GET("/swagger", func(c *gin.Context) {
		c.Redirect(http.StatusMovedPermanently, "/swagger/index.html")
	})
	r.GET("/swagger/index.html", func(c *gin.Context) {
		c.Header("Content-Type", "text/html")
		html := `<!DOCTYPE html>
<html lang="en">
<head>
  <meta charset="UTF-8">
  <title>API Documentation</title>
  <link rel="stylesheet" type="text/css" href="https://unpkg.com/swagger-ui-dist@5/swagger-ui.css" />
  <style>.topbar{display:none!important}</style>
</head>
<body>
  <div id="swagger-ui"></div>
  <script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js"></script>
  <script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-standalone-preset.js"></script>
  <script>
    window.onload=()=>{
      SwaggerUIBundle({
        url:"/swagger/doc.json",
        dom_id:'#swagger-ui',
        deepLinking:true,
        layout:"BaseLayout",
        presets:[SwaggerUIBundle.presets.apis,SwaggerUIStandalonePreset]
      });
    };
  </script>
</body>
</html>`
		c.String(http.StatusOK, html)
	})
	r.GET("/swagger/doc.json", func(c *gin.Context) {
		c.Header("Cache-Control", "no-cache, no-store, must-revalidate")
		c.File("./docs/swagger.json")
	})

	// Routes
	c := r.Group("/categories")
	{
		c.GET("", categoryHandler.GetAllCategories)
		c.POST("", categoryHandler.CreateCategory)
		c.GET("/:id", categoryHandler.GetCategoryByID)
		c.PUT("/:id", categoryHandler.UpdateCategory)
		c.DELETE("/:id", categoryHandler.DeleteCategory)
	}

	p := r.Group("/products")
	{
		p.GET("", productHandler.GetAllProducts)
		p.POST("", productHandler.CreateProduct)
		p.GET("/:id", productHandler.GetProductByID)
		p.PUT("/:id", productHandler.UpdateProduct)
		p.DELETE("/:id", productHandler.DeleteProduct)
	}

	api := r.Group("/api")
	{
		api.POST("/checkout", transactionHandler.Checkout)
		api.GET("/report/today", transactionHandler.ReportToday)
		api.GET("/report", transactionHandler.ReportRange)
	}

	// Redirect 404
	r.NoRoute(func(c *gin.Context) {
		path := c.Request.URL.Path
		// If path starts with /swagger/ but is not a file (no extension), redirect to index
		if strings.HasPrefix(path, "/swagger/") && !strings.Contains(path, ".") {
			c.Redirect(http.StatusFound, "/swagger/index.html")
		} else {
			// All other 404s go to root
			c.Redirect(http.StatusFound, "/")
		}
	})

	// Start server
	log.Println("Server running on port 8080")
	if err := r.Run(":8080"); err != nil {
		log.Fatal("Failed to start server: ", err)
	}
}